+++ clients/python/pycellbase/__init__.py
@@ -4,5 +4,5 @@ __credits__ = [""]
 __license__ = "Apache License 2.0"
 __version__ = "0.1"
 __maintainer__ = "Daniel Perez-Gil"
-__email__ = "daniel.perez@incliva.es"
+__email__ = "opencb@googlegroups.com"
 __status__ = "Development"
